% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bigramfrequency.R
\name{bigramfrequency}
\alias{bigramfrequency}
\title{bigramfrequency}
\usage{
bigramfrequency(df, colonne, slice = NA)
}
\arguments{
\item{df}{Your data frame}

\item{colonne}{Your column}

\item{slice}{Allows you to keep only a small number of observations, starting from the first. Default set to NA}
}
\value{
Return the bigram frequency of the column you chose
}
\description{
bigramfrequency
}
