% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_period.R
\name{choose_period}
\alias{choose_period}
\title{choose_period}
\usage{
choose_period(df, colonne, startday, endday, starttime = "00:00:00",
  endtime = "23:59:59", sort = FALSE)
}
\arguments{
\item{df}{Your data frame}

\item{colonne}{Your column}

\item{startday}{The first day of the period you want to select}

\item{endday}{The last day of the period you want to select}

\item{starttime}{The start time of the period you want to select. Default set to 00:00:00}

\item{endtime}{The end time of the period you want to select. Default set to 23:59:59}

\item{sort}{Gives the possibility to arrange the table, in ascending order}
}
\value{
Return the observations between the two specified dates
}
\description{
Time Filter function
}
