% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_fr.R
\name{dates_fr}
\alias{dates_fr}
\title{dates_fr}
\usage{
dates_fr(datum = Sys.time())
}
\arguments{
\item{datum}{Your date. DEFAULT set to Sys.time()}
}
\value{
Turn your date into a french format
}
\description{
dates_fr
}
