% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_fr.R
\name{datetimes_fr}
\alias{datetimes_fr}
\title{datetimes_fr}
\usage{
datetimes_fr(datum = Sys.time())
}
\arguments{
\item{datum}{Your datetime. DEFAULT set to Sys.time()}
}
\value{
Turn your datetime into a french format
}
\description{
datetimes_fr
}
