% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatlist.R
\name{flatlist}
\alias{flatlist}
\title{flatlist}
\usage{
flatlist(df)
}
\arguments{
\item{df}{Your data frame}
}
\value{
allows you to manage the columns of an array in list format, transforming them into character columns.
}
\description{
flatlist
}
