% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variable.R
\name{get_milk_data}
\alias{get_milk_data}
\title{Get Milk/Bottle Feeding Data}
\usage{
get_milk_data(data, arrange_by = "datetime")
}
\arguments{
\item{data}{A data frame containing the measurement data.}

\item{arrange_by}{Character string specifying column to arrange by. Default is "datetime".}
}
\value{
A filtered data frame containing only bottle feeding data.
}
\description{
Convenience function to filter for milk/bottle feeding measurements.
Equivalent to \code{filter_variable(data, "biberon")}.
}
\examples{
get_milk_data(example_data)
}
