% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variable.R
\name{get_size_data}
\alias{get_size_data}
\title{Get Size/Height Data}
\usage{
get_size_data(data, arrange_by = "datetime")
}
\arguments{
\item{data}{A data frame containing the measurement data.}

\item{arrange_by}{Character string specifying column to arrange by. Default is "datetime".}
}
\value{
A filtered data frame containing only size data.
}
\description{
Convenience function to filter for size/height measurements.
Equivalent to \code{filter_variable(data, "taille")}.
}
\examples{
get_size_data(example_data)
}
