% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_variable.R
\name{get_weight_data}
\alias{get_weight_data}
\title{Get Weight Data}
\usage{
get_weight_data(data, arrange_by = "datetime")
}
\arguments{
\item{data}{A data frame containing the measurement data.}

\item{arrange_by}{Character string specifying column to arrange by. Default is "datetime".}
}
\value{
A filtered data frame containing only weight data.
}
\description{
Convenience function to filter for weight measurements.
Equivalent to \code{filter_variable(data, "poids")}.
}
\examples{
get_weight_data(example_data)
}
