% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtagfrequency.R
\name{hashtagfrequency}
\alias{hashtagfrequency}
\title{hashtagfrequency}
\usage{
hashtagfrequency(df, colonne)
}
\arguments{
\item{df}{Your data frame}

\item{colonne}{Your column}
}
\value{
Return the hashtag frequency of the column you chose
}
\description{
hashtagfrequency
}
