% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Plot Timeline}
\usage{
plot_timeline(data)
}
\arguments{
\item{data}{A data frame containing health tracking data}
}
\value{
A ggplot object
}
\description{
Creates a vertical timeline visualization of events and symptoms.
}
