% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/povertext.R
\name{povertext}
\alias{povertext}
\title{povertext}
\usage{
povertext(val)
}
\arguments{
\item{val}{Your data frame}
}
\value{
Return a text, without any accent or links
}
\description{
Deplete the text value
}
