% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceByZero.R
\name{replaceByZero}
\alias{replaceByZero}
\title{replaceByZero}
\usage{
replaceByZero(df)
}
\arguments{
\item{df}{Your data frame}
}
\value{
Return a dataframe with no NA cells
}
\description{
replaceByZero
}
