% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_dates.R
\name{restore_dates}
\alias{restore_dates}
\title{restore_dates}
\usage{
restore_dates(column)
}
\arguments{
\item{column}{Your column}
}
\value{
Returns a column of datetimes compatible with R
}
\description{
restore_dates
}
