% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_empty_cols.R
\name{rm_empty_cols}
\alias{rm_empty_cols}
\title{rm_empty_cols}
\usage{
rm_empty_cols(df)
}
\arguments{
\item{df}{Your data frame}
}
\value{
Delete columns containing only NAs.
}
\description{
rm_empty_cols
}
