% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstats.R
\docType{data}
\name{rstats}
\alias{rstats}
\title{rstats}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 88 columns.
}
\usage{
data(rstats)
}
\description{
Description
}
\keyword{datasets}
