% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.R
\name{sampler}
\alias{sampler}
\title{sampler}
\usage{
sampler(df, nb = nrow(df))
}
\arguments{
\item{df}{Your data frame}

\item{nb}{Number of observations for the sample}
}
\value{
Return a sample of a dataframe
}
\description{
sampler
}
