% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wf_dictionary.R
\docType{data}
\name{wf_dictionary}
\alias{wf_dictionary}
\title{wf_dictionary}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 998 rows and 2 columns.}
\usage{
data(wf_dictionary)
}
\description{
Descripion
}
\keyword{datasets}
