% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordfrequency.R
\name{wordfrequency}
\alias{wordfrequency}
\title{wordfrenquency}
\usage{
wordfrequency(df, colonne)
}
\arguments{
\item{df}{Your data frame}

\item{colonne}{Your column}

\item{shortlist}{Slice the word frequency. Default set to no slice}
}
\value{
Return the word frequency of the column you chose
}
\description{
wordfrenquency
}
