% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_time.R
\name{distribution_time}
\alias{distribution_time}
\title{distribution_time}
\usage{
distribution_time(df, colonne)
}
\arguments{
\item{df}{Your data frame}

\item{colonne}{Your column}
}
\value{
Return a data frame with the distribution
}
\description{
Check the time distribution of a data frame
}
