% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtagfrequency.R
\name{hashtagfrequency}
\alias{hashtagfrequency}
\title{hashtagfrequency}
\usage{
hashtagfrequency(df, colonne, slice = NA)
}
\arguments{
\item{df}{Your data frame}

\item{colonne}{Your column}

\item{slice}{Allows you to keep only a small number of observations, starting from the first. Default set to NA}
}
\value{
Return the hashtag frequency of the column you chose
}
\description{
hashtagfrequency
}
