% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordfrequency.R
\name{wordfrequency}
\alias{wordfrequency}
\title{wordfrenquency}
\usage{
wordfrequency(df, colonne, slice = NA)
}
\arguments{
\item{df}{Your data frame}

\item{colonne}{Your column}

\item{slice}{Allows you to keep only a small number of observations, starting from the first. Default set to NA}
}
\value{
Return the word frequency of the column you chose
}
\description{
wordfrenquency
}
